+++ tools/cabana/mainwin.h
@@ -8,7 +8,7 @@
 #include <QSplitter>
 #include <QStatusBar>
 
-#include "tools/cabana/chartswidget.h"
+#include "tools/cabana/chart/chartswidget.h"
 #include "tools/cabana/detailwidget.h"
 #include "tools/cabana/messageswidget.h"
 #include "tools/cabana/videowidget.h"
